requirements.txt
aniso8601
Click
Flask
Flask-Cors
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
pytz
six
SQLAlchemy
Werkzeug
python-dotenv
gunicorn